Feat: Source buffer queueing
This allows streaming audio by buffering sound as it arrives, and
queueing said buffers onto the source. Processed buffers can be unqueued
and reused or deleted.

Just like with static buffers, cyal automatically manages the lifetime
of the Python Buffer objects, storing references to them when queued,
and returning these references to the library user when unqueued.