Feat: Implement AL_SOFT_deferred_updates extension
This is a very small extension which can be emulated on some platforms
with alc{Suspend,Process}Context(), so I implemented it directly in the
Context class.

Passing emulate_defer_updates=True (the default) to the Context
constructor will emulate the extension with
alc{Suspend,Process}Context(). Passing False will cause defer_updates()
and process_updates() to do nothing.

Primarily designed to be used via the Context.batch() @contextmanager
function.