Feat: Allow contexts to be made current upon construction
Specifying make_current=True in the Context constructor will cause the
context to be set as the current context. We have to (at least
temporarily) set the context as current when constructing it anyway, in
order to check for extensions, so setting this to True just stops us
from switching back to the previous context.