Feat: Use scikit-build-core + CMake as build system, Windows builds 🎉
Instead of using setuptools, we use scikit-build-core (a re-imagining of
the original scikit-build). This then defers to CMake, which eefinds the
system's OpenAL implementation, downloading and compiling the latest
OpenAL Soft if one could not be found. It also handles cythonizing and
compiling the Cyal extension modules themselves.

The fallback OpenAL Soft is compiled with minimal options (only the
library is built, none of the utils or examples are). The library is
installed to the "cyal" site-packages directory so it can be loaded by
the extension modules (this will likely need tweaking on Unix
platforms).