Feat: Implement exceptions.UnsupportedExtensionError
This is raised when a features available with an extension is requested,
but the OpenAL extension isn't supported by the implementation, and the
user hasn't requested that it be emulated by Cyal.

The option to emulate the simpler extensions is a convenience to
applications so that tey don't have to handle this exception for
features that aren't strictly necessary (E.G. if setting
Source.direct_channels = True fails, the application will still work,
but the quality of the audio will be diminished).