Feat: Implement AL_SOFT_DIRECT_CHANNELS extension
This is implemented as a Source property. The behaviour when the
implementation doesn't support the extension is governed by thhe
emulate_direct_channels kwarg passed to the Context constructor. If True
(the default), setting Source.direct_channels is silently ignored, and
requesting its value always returns False. If emulate_direct_channels is
False, an UnsupportedExtensionError() is raised if the property is
accessed or set.