Feat: Implement alcResetDeviceSOFT
Very similar to alcReopenDeviceSOFT but without the device name, call
Device.reset() to use it.

This is a weird one since it's defined in two extensions: ALC_SOFT_HRTF
and ALC_SOFT_OUTPUT_MODE, and isn't defined in its own extension (like
it probably should be) so the error you get if it isn't available may be
a little strange.